+++ surprise/dump.py
@@ -3,63 +3,55 @@ The :mod:`dump` module defines the :func:`dump` function.
 """
 
 import pickle
-import importlib
 
 
-def dump(file_name, predictions, trainset=None, algo=None):
-    """Dump a list of :obj:`predictions
-    <surprise.prediction_algorithms.predictions.Prediction>` for future
-    analysis, using Pickle.
+def dump(file_name, predictions=None, algo=None, verbose=0):
+    """A basic wrapper around Pickle to serialize a list of prediction and/or
+    an algorithm on drive.
 
-    If needed, the :class:`trainset <surprise.dataset.Trainset>` object and the
-    algorithm can also be dumped. What is dumped is a dictionary with keys
-    ``'predictions``, ``'trainset'``, and ``'algo'``.
-
-    The dumped algorithm won't be a proper :class:`algorithm
-    <surprise.prediction_algorithms.algo_base.AlgoBase>` object but simply a
-    dictionary with the algorithm attributes as keys-values (technically, the
-    ``algo.__dict__`` attribute).
-
-    See :ref:`User Guide <dumping>` for usage.
+    What is dumped is a dictionary with keys ``'predictions'`` and ``'algo'``.
 
     Args:
         file_name(str): The name (with full path) specifying where to dump the
             predictions.
-
         predictions(list of :obj:`Prediction\
             <surprise.prediction_algorithms.predictions.Prediction>`): The
             predictions to dump.
-        trainset(:class:`Trainset <surprise.dataset.Trainset>`, optional): The
-            trainset to dump.
         algo(:class:`Algorithm\
             <surprise.prediction_algorithms.algo_base.AlgoBase>`, optional):
-            algorithm to dump.
+            The algorithm to dump.
+        verbose(int): Level of verbosity. If ``1``, then a message indicates
+            that the dumping went succesfully. Default is ``0``.
     """
 
-    dump_obj = dict()
-
-    dump_obj['predictions'] = predictions
-
-    if trainset is not None:
-        dump_obj['trainset'] = trainset
-
-    if algo is not None:
-        dump_obj['algo'] = algo.__dict__  # add algo attributes
-        dump_obj['algo']['name'] = algo.__class__.__name__
-
+    dump_obj = {'predictions': predictions,
+                'algo': algo
+                }
     pickle.dump(dump_obj, open(file_name, 'wb'))
-    print('The dump has been saved as file', file_name)
+
+    if verbose:
+        print('The dump has been saved as file', file_name)
 
 
-def load_algo(file_name):
-    """Load a prediction algorithm from a file, using Pickle.
+def load(file_name):
+    """A basic wrapper around Pickle to deserialize a list of prediction and/or
+    an algorithm that were dumped on drive using :func:`dump()
+    <surprise.dump.dump>`.
 
     Args:
         file_name(str): The path of the file from which the algorithm is
             to be loaded
+
+    Returns:
+        A tuple ``(predictions, algo)`` where ``predictions`` is a list of
+        :class:`Prediction
+        <surprise.prediction_algorithms.predictions.Prediction>` objects and
+        ``algo`` is an :class:`Algorithm
+        <surprise.prediction_algorithms.algo_base.AlgoBase>` object. Depending
+        on what was dumped, some of these may be ``None``.
+
     """
+
     dump_obj = pickle.load(open(file_name, 'rb'))
-    algo_module = importlib.import_module('surprise.prediction_algorithms')
-    algo = getattr(algo_module, dump_obj['algo']['name'])()
-    algo.__dict__ = dump_obj['algo']
-    return algo
+
+    return dump_obj['predictions'], dump_obj['algo']